rename file names from pretixXXX to zuzaluPretixXXX
