+++ apps/passport-server/src/services.ts
@@ -6,13 +6,13 @@ import { startEmailTokenService } from "./services/emailTokenService";
 import { startIssuanceService } from "./services/issuanceService";
 import { startMetricsService } from "./services/metricsService";
 import { startPersistentCacheService } from "./services/persistentCacheService";
-import { startPretixSyncService } from "./services/pretixSyncService";
 import { startProvingService } from "./services/provingService";
 import { startRollbarService } from "./services/rollbarService";
 import { startSemaphoreService } from "./services/semaphoreService";
 import { startTelegramService } from "./services/telegramService";
 import { startTelemetry } from "./services/telemetryService";
 import { startUserService } from "./services/userService";
+import { startZuzaluPretixSyncService } from "./services/zuzaluPretixSyncService";
 import { APIs, ApplicationContext, GlobalServices } from "./types";
 
 export async function startServices(
@@ -27,7 +27,7 @@ export async function startServices(
   const emailService = startEmailService(context, apis.emailAPI);
   const emailTokenService = startEmailTokenService(context);
   const semaphoreService = startSemaphoreService(context);
-  const zuzaluPretixSyncService = startPretixSyncService(
+  const zuzaluPretixSyncService = startZuzaluPretixSyncService(
     context,
     rollbarService,
     semaphoreService,
